create client class as the data source
